LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.measure.baselines-expected.txt

2d.text.measure.baselines
Testing baselines
Actual output:

FAIL Testing baselines assert_equals: ctx.measureText('A').ideographicBaseline === 6.25 (got -13[number], expected 6.25[number]) expected 6.25 but got -13
